めちゃくちゃリファクタした
